资源校验

+++ app/src/main/java/io/shulie/takin/cloud/app/entity/JobEntity.java
@@ -30,7 +30,7 @@ public class JobEntity {
     /**
      * 任务持续时长
      */
-    private Long duration;
+    private Integer duration;
     /**
      * 采样率
      */
